Readcount to be corrected
Readcount still implemented if not required by policy. TODO: add or
remove cog such that only one cog is implemented at a time
