+++ include/cog.h
@@ -1,18 +1,7 @@
 #ifndef COG_H_
-#define COG_H_
-
-class cog {
-	protected:
-		const int type;
-		unsigned int readcount;
-		
-	public:
-		cog (int t);
-		const int getType();
-		
-		cog (int t, unsigned int r);
-		unsigned int getReadcount();
-		void setReadcount(unsigned int x);
-};
-
-#endif /* COG_H_ */
+#ifdef READCOUNT
+#include "cogWithRead.h"
+#else
+#include "cogWithoutRead.cpp"
+#endif
+#endif
