+++ include/cogWithoutRead.h
@@ -1,0 +1,15 @@
+#ifndef COG_H_
+#ifndef READCOUNT
+#define COG_H_
+
+class cog {
+	protected:
+		const int type;
+		
+	public:
+		cog (int t);
+		const int getType();
+};
+
+#endif
+#endif
