+++ include/policyTypes.h
@@ -5,8 +5,12 @@
 #include "crackPolicy.cpp"
 #include "sortPolicy.h"
 #include "sortPolicy.cpp"
+
+#ifdef READCOUNT
 #include "splayPolicy.h"
 #include "splayPolicy.cpp"
+#endif
+
 #include "hybridPolicy.h"
 #include "hybridPolicy.cpp"
