+++ include/printTree.h
@@ -1,0 +1,8 @@
+#ifndef PRINTTREE_H_
+#define PRINTTREE_H_
+
+#include "cog.h"
+
+void printTree(cog* it,int depth);
+
+#endif /* PRINTTREE_H_ */
