+++ include/splayPolicy.h
@@ -1,4 +1,5 @@
 #ifndef SPLAYPOLICY_H_
+#ifdef READCOUNT
 #define SPLAYPOLICY_H_
 
 #include "cog.h"
@@ -14,4 +15,5 @@ class splayPolicy: public policy<T> {
 		void afterIterator	(cog* &node);
 };
 
+#endif
 #endif /* SPLAYPOLICY_H_ */
